Type-cast to int instead of float during binarizing the images

+++ src/metrics.py
@@ -22,6 +22,13 @@ def calculate_dice_similairity(seg,gt):
     seg = seg.view(-1)
     gt = gt.view(-1)
 
+
+    # Rationale behind the conversion to numpy (and the evenetual migration of computation to the CPU):
+    #   - Binarizing the tensor by applying the conditional casts it to a np ndarray of bool type
+    #   - This np ndarray can't be cast back to a PyTorch tensor
+    #   - Cleaner to convert the torch.Tensor to numpy.ndarray at the start
+    #   - Through the type-check, we retain the flexibility to supply either a numpy array or a torch.Tensor as the input arguments
+
     if torch.is_tensor(seg) is True:
         seg = seg.detach()
         seg = seg.cpu().numpy()
@@ -31,8 +38,8 @@ def calculate_dice_similairity(seg,gt):
         gt = gt.cpu().numpy()
 
     #Binarize the images
-    seg = (seg > 0.5).astype(float)
-    gt = (gt > 0.5).astype(float)
+    seg = (seg > 0.5).astype(int)
+    gt = (gt > 0.5).astype(int)
 
     inter = np.inner(seg,gt)
     union = np.sum(seg) + np.sum(gt)
